Move BacklogsPage logic to view model
